pyproject.toml
[project]
name = "gauntlet"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
]

[project.scripts]
gauntlet-mcp = "gauntlet.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
markers = ["docker: requires Docker and docker compose (deselect with -m 'not docker')"]
addopts = "--cov=gauntlet --cov-report=term-missing --cov-report=xml"
norecursedirs = [".gauntlet"]

[tool.coverage.run]
branch = true

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B"]

[tool.mypy]
strict = true
python_version = "3.10"

[dependency-groups]
dev = [
    "hatchling>=1.29.0",
    "mypy>=1.20.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.10",
    "types-pyyaml>=6.0.12.20260408",
    "types-requests>=2.33.0.20260408",
]
